+++ arm9/source/patches/arm9/sdk5/CardiReadRomWithCpuPatch.h
@@ -8,6 +8,14 @@ public:
     bool FindPatchTarget(PatchContext& patchContext) override;
     void ApplyPatch(PatchContext& patchContext) override;
 
+    /// @brief Returns the location of CARDi_ReadRomWithCPU. Only valid after FindPatchTarget.
+    /// @return The location of CARDi_ReadRomWithCPU, or \c nullptr if it was not found.
+    constexpr const u32* GetCardiReadRomWithCpu() const { return _cardiReadRomWithCpu; }
+
+    /// @brief Returns whether CARDi_ReadRomWithCPU is Thumb code. Only valid after FindPatchTarget.
+    /// @return True if CARDi_ReadRomWithCPU is Thumb code, false if it is ARM code.
+    constexpr bool IsThumb() const { return _thumb; }
+
 private:
     u32* _cardiReadRomWithCpu = nullptr;
     u16 _thumb = false;