+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Support for banner save files
 - In-game reset: hold L+R+Down+B for two seconds in a retail game to return to the launcher
 
+### Fixed
+- Boot hang in Digging for Dinosaurs, The Magic School Bus: Oceans, I Spy Game Pack, I Spy Universe, 2 Game Pack: My Amusement Park + Digging for Dinosaurs and Animal Planet: Vet Collection - by @danyocom
+
 ## [v1.7.1] - 28 Jun 2026
 
 ### Fixed